pyproject.toml
[project]
name = "simple-ansible-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "pyyaml>=6.0.3",
]
